update model dan migration

+++ database/migrations/2025_10_13_041447_create_movies_table.php
@@ -6,17 +6,23 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
+    /**
+     * Jalankan migrasi.
+     */
     public function up(): void
     {
         Schema::create('movies', function (Blueprint $table) {
             $table->id();
-            $table->string('judul');
-            $table->string('genre');
-            $table->float('rating')->default(0);
-            $table->timestamps();
+            $table->string('judul');           
+            $table->string('genre');             
+            $table->decimal('rating', 3, 1)->default(0); 
+            $table->timestamps();                
         });
     }
 
+    /**
+     * Kembalikan (rollback) migrasi.
+     */
     public function down(): void
     {
         Schema::dropIfExists('movies');
